+++ .eslintrc.yml
@@ -1,4 +1,5 @@
 parserOptions:
+  ecmaVersion: 2015
   sourceType: 'module'
 
 env:
