+++ src/components/tripRouteTemplate.js
@@ -1,0 +1,11 @@
+/*
+* возвращает шаблон верстки маршрута
+*/
+const createTripRouteTemplate = () =>
+  `<div class="trip-info__main">
+    <h1 class="trip-info__title">Amsterdam &mdash; ... &mdash; Amsterdam</h1>
+
+    <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;21</p>
+  </div>`;
+
+export {createTripRouteTemplate};
